requirements.txt
DateTime==5.0
pandas==1.5.3
python-dotenv==0.21.1
requests==2.28.2
